requirements.txt
# Core ML
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.0
imbalanced-learn==0.12.2
# Visualization
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0
# LLM + Agents (Groq - free API)
groq==0.37.1
langchain-groq==1.1.2
langchain-core==1.2.17
langchain==0.3.25
langchain-community==0.3.24
langsmith==0.7.13
# Dashboard
streamlit==1.35.0
# Utilities
python-dotenv==1.0.1
joblib==1.4.2
notebook==7.2.0
requests==2.32.5
